final version

+++ Assets/Scripts/HealthBar.cs
@@ -1,11 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 public class HealthBar : MonoBehaviour
 {
     [SerializeField] private Slider slider;
-    public void UpdateHealthBar(int health)
+    public void UpdateHealthBar(float health)
     {
         if (health <= 0) 
         {
